requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6